+++ eidolon/Sources/SwiftUI/Core.swift
@@ -2,11 +2,7 @@ import UIKit
 @_exported import CoreGraphics
 @_exported import UIKit
 @_exported import Foundation
-#if REV_SYSTEM_COMBINE || REV_STYX
 @_exported import Combine
-#else
-@_exported import OpenCombine
-#endif
 
 @MainActor @preconcurrency public protocol View {
     associatedtype Body: View